% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/northarrow.R
\name{addnortharrow}
\alias{addnortharrow}
\title{Plot North Arrow}
\usage{
addnortharrow(pos = "topright", padin = c(0.15, 0.15), scale = 1,
  lwd = 1, border = "black", cols = c("white", "black"),
  text.col = "black")
}
\arguments{
\item{pos}{Where to align the north arrow. One of "bottomleft", "bottomright", "topleft",
or "topright".}

\item{padin}{A vector of length 2 determining the distance in inches between the scalebar
and the edge of the plottable area.}

\item{scale}{Scale the default north arrow to make it bigger or smaller}

\item{lwd}{The line width outlining the north arrow}

\item{border}{The line color outlining the north arrow}

\item{cols}{A vector of length 2 determining the two colors to be drawn for the north arrow}

\item{text.col}{Color of the "N"}
}
\description{
Plot a north arrow (pointing directly "up") positioned based on current
plot extents.
}
\examples{
\donttest{
library(maptools)
data(wrld_simpl)
plot(wrld_simpl)
addnortharrow()
}
plot(1:5, 1:5, asp=1)
addnortharrow()
}

