% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caches.R
\name{clear_geocode_cache}
\alias{clear_geocode_cache}
\title{Clear cached results}
\usage{
clear_geocode_cache(cache = NA, ...)
}
\arguments{
\item{cache}{A folder name where cached files should be stored.}
}
\description{
Clears the local cache of downloaded files (by default in the
rest.cache folder in the working directory).
}
\examples{
clear_rest_cache()

}

