% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{get_default_geocoder}
\alias{get_default_geocoder}
\alias{set_default_geocoder}
\title{Get/Set the default geocoder}
\usage{
get_default_geocoder()

set_default_geocoder(geocoder)
}
\arguments{
\item{geocoder}{The new source to use. One of "pickpoint", "google", or "dsk".}
}
\description{
The \link{geocode} function can use google, pickpoint, or data science toolkit
to turn human-readable names into coordinates. Use these methods to get/set
the default source. These will need to be called once per namespace load.
}
\examples{
get_default_geocoder()
set_default_geocoder("google")
(set_default_geocoder(NULL))

}

