% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bboxtools.R
\name{makebbox}
\alias{makebbox}
\title{Create a Bounding Box}
\usage{
makebbox(n, e, s, w)
}
\arguments{
\item{n}{North bounding latitude}

\item{e}{East bounding longitude}

\item{s}{South bounding latitude}

\item{w}{West bounding longitude}
}
\value{
A 2x2 matrix describing a bounding box like that returned by \code{sp::bbox()}
}
\description{
Convencience method to create a bounding box like that returned by \code{sp::bbox()}.
To generate a bounding box from lists of lat/lon values use \code{sp::bbox(cbind(lons, lats))}.
}
\examples{
makebbox(45.125, -64.25, 44.875, -64.75)
}
\seealso{
sp::bbox
}

