% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxtools.R
\name{mergebbox}
\alias{mergebbox}
\title{Combine bounding boxes}
\usage{
mergebbox(...)
}
\arguments{
\item{...}{An arbitrary number of bounding boxes as generated by \code{sp::bbox},
\link{makebbox} or \link{searchbbox}}
}
\value{
A single bounding box that contains all of its arguments.
}
\description{
Create a single bounding box that encloses all of the bounding boxes.
}
\examples{

box1 <- makebbox(45, -64, 44, -65)
box2 <- makebbox(45.5, -64.5, 44.5, -65.6)
mergebbox(box1, box2)

}
