% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prettymap.R
\name{prettymap}
\alias{prettymap}
\title{Plot A Pretty Map}
\usage{
prettymap(plotexpression, oma = c(0, 0, 0, 0), mai = c(0.1, 0.1, 0.1, 0.1),
  drawbox = TRUE, box.lwd = 1, drawscale = TRUE,
  scale.pos = "bottomleft", scale.htin = 0.1, scale.widthhint = 0.25,
  scale.unitcategory = "metric", scale.style = "bar",
  scale.bar.cols = c("black", "white"), scale.lwd = 1,
  scale.linecol = "black", scale.padin = c(0.15, 0.15),
  scale.labelpadin = 0.08, scale.label.cex = 0.8,
  scale.label.col = "black", scale.plotunit = NULL, scale.plotepsg = NULL)
}
\arguments{
\item{plotexpression}{}

\item{oma}{}

\item{mai}{}

\item{drawbox}{}

\item{box.lwd}{}

\item{drawscale}{}

\item{scale.pos}{}

\item{scale.htin}{}

\item{scale.widthhint}{}

\item{scale.unitcategory}{}

\item{scale.style}{}

\item{scale.bar.cols}{}

\item{scale.lwd}{}

\item{scale.linecol}{}

\item{scale.padin}{}

\item{scale.labelpadin}{}

\item{scale.label.cex}{}

\item{scale.label.col}{}

\item{scale.plotunit}{}

\item{scale.plotepsg}{}
}
\description{
Plot A Pretty Map
}

