% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bboxtools.R
\name{searchbbox}
\alias{searchbbox}
\title{Query The Interwebs For A Bounding Box}
\usage{
searchbbox(querystring, source = "dsk")
}
\arguments{
\item{querystring}{The search query}

\item{source}{One of \code{dsk} or {google}}
}
\value{
A 2x2 matrix describing a bounding box like that returned by \code{sp::bbox()}
}
\description{
Use the Data Science Toolkit (\url{http://www.datasciencetoolkit.org/about}) to
retreive a bounding box for the given query. Implemented
from the \code{ggmap:geocode} function from the \code{ggmap}
package (\url{https://cran.r-project.org/package=ggmap})
by David Kahle to remove dependencies
of \code{ggmap} that are not necessary for \code{prettymapr}.
}
\examples{
\donttest{
searchbbox("kings county, NS")
searchbbox("University Ave. Wolfville NS")
searchbbox("Wolfville ns")
}
}

