% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caches.R
\name{set_cached}
\alias{set_cached}
\alias{get_cached}
\alias{clear_cache}
\alias{cache_size}
\title{Internal cache methods}
\usage{
set_cached(cache, url, data, ...)

get_cached(cache, url, ...)

clear_cache(cache, ...)

cache_size(cache, ...)
}
\arguments{
\item{cache, url, data, ...}{For internal use only}
}
\value{
Values for internal use
}
\description{
For internal use only.
}
