% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caches.R
\name{clear_geocode_cache}
\alias{clear_geocode_cache}
\title{Clear cached results}
\usage{
clear_geocode_cache(cache = NA)
}
\arguments{
\item{cache}{An environment, a directory name, or NA to clear
the default internal cache}
}
\description{
Clears the local cache of downloaded files (by default, an
environment in the package namespace). Clearing a directory
cache will result in all files with the extention ".cached"
being deleted from that directory.
}
\examples{
clear_geocode_cache()

}
