% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxtools.R
\name{mergebbox}
\alias{mergebbox}
\title{Combine bounding boxes}
\usage{
mergebbox(...)
}
\arguments{
\item{...}{An arbitrary number of bounding boxes as generated by \code{sp::bbox},
\link{makebbox} or \link{searchbbox}}
}
\value{
A single bounding box that contains all of its arguments.
}
\description{
Create a single bounding box that encloses each bounding box.
}
\examples{
library(prettymapr)
box1 <- searchbbox("wolfville, ns")
box2 <- searchbbox("halifax, ns")
box3 <- searchbbox("kentville, ns")
mergebbox(box1, box2, box3)

}

