% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bboxtools.R
\name{searchbbox}
\alias{searchbbox}
\title{Query The Interwebs For A Bounding Box}
\usage{
searchbbox(querystring, ...)
}
\arguments{
\item{querystring}{The search query.}

\item{...}{Additional paramters to be passed on to \link{geocode}. Passing \code{source="google"}
may be useful if google is desired as a source. Use \code{options(prettymapr.geosource="google")}
to permanently use \code{google} as a source.}
}
\value{
A 2x2 matrix describing a bounding box like that returned by \code{sp::bbox()}
}
\description{
Use the \href{http://www.pickpoint.io/}{PickPoint.io API} to
retreive a bounding box for the given query. Implemented
from the \code{ggmap:geocode} function from the \code{ggmap}
package (\url{https://cran.r-project.org/package=ggmap})
by David Kahle to remove dependencies
of \code{ggmap} that are not necessary for \code{prettymapr}. Note that if
you would like to use \code{google} as a source, you must agree to the Google
API \href{https://developers.google.com/terms/}{terms and conditions}.
}
\examples{
#don't test to speed up checking time
\donttest{
searchbbox("kings county, NS")
searchbbox("University Ave. Wolfville NS", source="google")
searchbbox("Wolfville ns", source="google")
searchbbox(c("Vermont", "Nova Scotia"))
}
}

